Algorithm separated

+++ Constants.java
@@ -25,7 +25,7 @@ public class Constants {
 	public static int MAX_TRANS = 5;
 	public final static double nIdleTarget = 0;
 
-	public static 		double EPS = 6.0;
+	public static 		int EPS = 6;
 	public static   	double BETA = 0.75;
 	public static   	double GEMMA = 4.0;
 	public static 		double ALPHA = 0.9375;
